+++ src/views/productView.js
@@ -14,11 +14,16 @@ const errors = {
       err.statusCode = 400;
       return err;
     })(),
+    idNotFound: (() => {
+        const err = Error("Id not found");
+        err.statusCode = 404;
+        return err;
+    })(),
   }
 
 module.exports = {    
     getAllProduct: async(req, res) => {
-        var elements = [];
+        let elements = [];
         const products = await Product.find();
         for(let i = 0; i < products.length; i++) {
             const article = await Article.findById(products[i].articleId);
@@ -41,25 +46,28 @@ module.exports = {
         if (!isValidObjectId(productId)) return errors.invalidId;
 
         const product = await Product.findById(productId);
-        const article = await Article.findById(product.articleId);
-        var element  = { 
-            articleId: article.id, 
-            productId: product.id, 
-            name: article.name, 
-            price: article.price, 
-            description: article.description, 
-            restaurantId: article.restaurantId, 
-            imageUrl: article.imageUrl, 
-            allergenList: product.allergenList, 
-            ingredientList: product.ingredientList 
+        if(product) {
+            const article = await Article.findById(product.articleId);
+            let element  = { 
+                articleId: article.id, 
+                productId: product.id, 
+                name: article.name, 
+                price: article.price, 
+                description: article.description, 
+                restaurantId: article.restaurantId, 
+                imageUrl: article.imageUrl, 
+                allergenList: product.allergenList, 
+                ingredientList: product.ingredientList 
+            }
+            return element;
         }
-        return element;
+        else return errors.idNotFound;
     },
     getProductByRestaurantId: async(req, res) => {
         const { restaurantId } = req.params;
         if (!isValidObjectId(restaurantId)) return errors.invalidId;
 
-        var elements = [];
+        let elements = [];
         const articles = await Article.find({restaurantId: restaurantId});
         for(let i = 0; i < articles.length; i++) {
             const product = await Product.findOne({articleId: articles[i].id});
@@ -75,14 +83,15 @@ module.exports = {
                     allergenList: product.allergenList, 
                     ingredientList: product.ingredientList 
                 })
+                return elements;
             }
+            else return errors.idNotFound;
         }
-        return elements;
     },
     getProductByName: async(req, res) => {
         const { productName } = req.params;
 
-        var elements = [];
+        let elements = [];
         const articles = await Article.find({name: { '$regex' : productName, '$options' : 'i' }});
         for(let i = 0; i < articles.length; i++) {
             const product = await Product.findOne({articleId: articles[i].id});
@@ -98,27 +107,40 @@ module.exports = {
                     allergenList: product.allergenList, 
                     ingredientList: product.ingredientList 
                 })
+                return elements;
             }
-        }
-        return elements;
+            else return errors.idNotFound;
+        } 
     },
     createProduct: async(req, res) => {
         const { name, price, description, restaurantId, imageUrl, allergenList, ingredientList } = req.body;
         if (!isValidObjectId(restaurantId)) return errors.invalidId;
 
-        const article = await Article.create({ name: name, price: price, description: description, restaurantId: restaurantId, imageUrl: imageUrl });
-        const resultProduct = await Product.create({ articleId: article.id, allergenList, ingredientList });
-        return `product ${resultProduct.id} created`;
+        const article = await Article.create({ 
+            name: name, 
+            price: price, 
+            description: description, 
+            restaurantId: restaurantId, 
+            imageUrl: imageUrl });
+
+        const product = await Product.create({ 
+            articleId: article.id, 
+            allergenList: allergenList, 
+            allergenList: ingredientList });
+
+        return `product ${product.id} created`;
     },
     deleteProduct: async(req, res) => {
         const { productId } = req.params;
         if (!isValidObjectId(productId)) return errors.invalidId;
 
         // delete the product
-        const product = await Product.findById(productId);
-        await Article.findByIdAndDelete(product.articleId);
-        await Product.findByIdAndDelete(productId);
-        
+        const product = await Product.findByIdAndDelete(productId);
+        if(product) {
+            await Article.findByIdAndDelete(product.articleId);
+        }
+        else return errors.idNotFound;
+
         // delete in menus
         await Menu.updateMany({ }, { $pull: { productIdList: { $in: [ productId ]} } })
 
@@ -137,15 +159,18 @@ module.exports = {
             ingredientList: ingredientList 
         });
 
-        await Article.findOneAndUpdate({
-            _id: product.articleId
-        }, {
-            name: name, 
-            price: price, 
-            description: description, 
-            restaurantId: restaurantId, 
-            imageUrl: imageUrl
-        });
+        if(product) {
+            await Article.findOneAndUpdate({
+                _id: product.articleId
+            }, {
+                name: name, 
+                price: price, 
+                description: description, 
+                restaurantId: restaurantId, 
+                imageUrl: imageUrl
+            });
+        }
+        else return errors.idNotFound;
 
         return `product ${productId} updated`;
     }
